refactor(coworker): extract route-context providers into page-owned modules (Phase 2 PR 2)
Pure mechanical move: the ~15 inline providers in route-context.ts are extracted
byte-for-byte into 7 domain modules under route-context/providers/ (compliance,
platform, ops, workspace, build, finance, commerce). No behavior change — the
golden route-context tests (exact-output assertions) stay green (16/16, incl.
#4051's inheritance conformance guard). route-context.ts shrinks 1117 -> 175 LOC
(now a thin registry + resolver + business-context + default label) and drops out
of the module-size >800 baseline. The registry imports each provider, so a page
self-registers by adding its module + a registry entry.

## Design grounding
- Existing specs/plans reviewed: docs/superpowers/specs/2026-08-05-coworker-page-perception-context-contract-design.md; docs/superpowers/plans/2026-08-06-page-owned-context-contract-phase2-plan.md
- Current code substrate reviewed: apps/web/lib/tak/route-context.ts; apps/web/lib/tak/route-context/providers/*.ts; apps/web/lib/tak/route-context-map.ts (peer injector #2 guarded by #4051)
- Source of truth: the design spec (Option B / Phase 2) plus the merged contract PR #4057
- Decision: mechanical page-owned-module extraction (PR 2 of Phase 2); no contract or behavior change; Phase 3 (unify injector #2 + surface->tool parity + kernel gate) remains follow-up

Docs-Impact-Decision: pure internal refactor moving provider functions between files; no change to the documented decision-perspective workflow or user-facing capability, so no user-guide page changes.
