+++ scripts/module-size-baseline.txt
@@ -58,7 +58,6 @@ apps/web/lib/tak/agent-routing.ts	1027
 apps/web/lib/tak/agentic-loop.test.ts	2387
 apps/web/lib/tak/agentic-loop.ts	2737
 apps/web/lib/tak/route-context-map.ts	1260
-apps/web/lib/tak/route-context.ts	1118
 apps/web/lib/work-capsules/mcp-handlers.ts	884
 apps/web/lib/work-capsules/work-capsule-store.test.ts	1103
 apps/web/lib/work-capsules/work-capsule-store.ts	1056
